+++ src/tinydiffusion/utils/__init__.py
@@ -6,6 +6,7 @@ from tinydiffusion.utils.tracking import (
     RunLogger,
     null_logger,
     quartile_means,
+    read_metrics,
     timestep_quartile_losses,
     timestep_quartile_totals,
 )
@@ -15,6 +16,7 @@ __all__ = [
     "eval_mode",
     "null_logger",
     "quartile_means",
+    "read_metrics",
     "seed_everything",
     "timestep_quartile_losses",
     "timestep_quartile_totals",